fix(testing): refactor kv index tests into testing package and integrate with bolt
